+++ db.json
@@ -10,8 +10,8 @@
     {
       "name": "아소비 서울길음롯데캐슬클라시아2호점",
       "defaultImgUrl": "https://search.pstatic.net/common/?type=b150&src=http://imgnews.naver.net/image/5239/2019/05/20/0000195912_001_20190520085409372.jpg",
-      "latitude": 1270264290,
-      "longitude": 376085003,
+      "latitude": 1270271290,
+      "longitude": 376089003,
       "averageRating": 0.0
     },
     {
